Cambios en la lectura de archivos para permitir instancias reales

+++ src/busqueda-local/operador_1.h
@@ -4,7 +4,7 @@
 #include "../utils/gap_instance.h"
 #include "../utils/solution.h"
 
-Solution operador_1(const GAPInstance& instance, int cmax);
+Solution operador_1(const GAPInstance& instance, double cmax);
 
 bool swap_op(Solution& solucion, const GAPInstance& instance, int j1, int j2);
