+++ src/busqueda-local/operador_2.h
@@ -4,7 +4,7 @@
 #include "../utils/solution.h"
 #include "../utils/gap_instance.h"
 
-Solution operador_2(const GAPInstance& instance, int cmax);
+Solution operador_2(const GAPInstance& instance, double cmax);
 
 bool relocate_op (Solution &solucion, const GAPInstance& instance, int i, int j);
