+++ src/main.cpp
@@ -30,7 +30,7 @@ int main(int argc,char** argv){
         std::vector<int> M(instance.m);
         for(int i=0; i<instance.n; i++) N[i]=i;
         for(int i=0; i<instance.m; i++) M[i]=i;
-        int cmax = costo_max(N,M,instance.costos);
+        double cmax = costo_max(N,M,instance.costos);
         auto solution = run_algorithm(algorithm,instance,cmax);
         save_solution(output_file, input_file, algorithm, solution, instance);
